DSPBRIDGE: Change dev_type to u8
dev_type can take one of the 2 values
DSP_UNIT    0
IVA_UNIT    1
So it need not be a s32 or u32. Hence changing it to u8.

+++ arch/arm/plat-omap/include/dspbridge/dev.h
@@ -351,7 +351,7 @@ extern dsp_status dev_get_dev_node(struct dev_object *hdev_obj,
  *      else:           *phDevNode is NULL.
  */
 extern dsp_status dev_get_dev_type(struct dev_object *hdevObject,
-				   u32 *dev_type);
+					u8 *dev_type);
 
 /*
  *  ======== dev_get_first ========
